Add files via upload
Danny's Diner, the owner of a small but popular diner, is looking to analyze the performance and operations of his diner using data to drive decisions. He has collected a variety of data points related to customer orders, menu items, and the membership status of his customers. Danny aims to understand customer behavior, menu performance, and the impact of the membership program on sales.